requirements.txt
supabase
fastapi
uvicorn
pydantic
langchain
openai
tiktoken
unstructured
slack_sdk
sentence-transformers
helicone
shortuuid
pandas
pinecone-client
slack_sdk
langdetect
pycld2
orjson